+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - #10006 Add fallback for /apps/default_id. Returns preconfigured id when User Profile Api responses with an error.
 
 #### Changed
+- #10095 Optimize EventShapeLayerProvider
 - #10164 Invoke layers providers in parallel
 
 #### Removed